:truck: Renamed and moved file managers to dao folder. Also added models folder
